refactor: multiline tree display
